# src/main/java/com/example/bankcards/dto/response/JwtResponse.java
package com.example.bankcards.dto.response;

import lombok.Builder;
import lombok.Data;

@Data
@Builder
public class JwtResponse {
    private String token;
    @Builder.Default
    private String type = "Bearer";
    private Long id;
    private String username;
    private String role;
}